Extracted signal power strategy
